% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihazr.R
\name{ihazr}
\alias{ihazr}
\title{ihazr}
\usage{
ihazr(time, status, marker, buttons = TRUE, binMax = 10,
  width = NULL, height = NULL)
}
\arguments{
\item{time}{A numerical vector of time data.}

\item{status}{A binary vector representing event or no event.}

\item{marker}{A collection of numerical variables of interest for survival
analysis.}

\item{buttons}{Boolean argument to decide if buttons or a drop down feature
will be used for variable selection. The default is set to TRUE.}

\item{binMax}{The maximum bin size for non parametric Epinechnikov hazard function.
This will only need to be changed for visualizing the kernel.}

\item{width}{The width of the html widget. The default is NULL, which results
in intelligent automatic sizing based on the widget's container.}

\item{height}{The height of the html widget. The default is NULL, which
results in intelligent automatic sizing based on the widget's container.}
}
\value{
ihazr interactive session for data values \code{time}, \code{status}, and \code{marker}
}
\description{
\code{ihazr} interactively updates a hazard function for survival anaylisis.
This function creates an HTML widget for survival data. The widget allows the
user to choose subsections of their data and plot the nonparametric kernel
estimated hazard function and cumulative hazard function. Bin size selection
for the Epinechnikov kernel is also made available for the user.
}
\examples{
Example 1 - simulated data
time_t <- runif(50, 0, 10)
status_t <- rbinom(50, 1, .7)
age_t <- runif(50, 20, 80)
ihazr(time_t, status_t, age_t)

Example 2 - with survival data
library(survival)
library(dplyr)
pbc5 <- pbc \%>\%
  slice(1:312) \%>\%
  select(time, status, age, edema, bili, albumin, protime) \%>\%
  mutate(time = time/365, status = (status==2)*1, bili = log(bili),
    protime = log(protime))
ihazr(time=pbc5[,1], status=pbc5[,2], marker=pbc5[,3:7])

Example 3 - in a Shiny app
library(shiny)
library(survival)
library(dplyr)
pbc5 <- pbc \%>\% slice(1:312) \%>\%
      select(time, status, age, edema, bili, albumin, protime) \%>\%
      mutate(time = time/365, status = (status==2)*1, bili = log(bili),
        protime = log(protime))
shinyApp(
  ui = ihazrOutput("test"),
  server = function(input, output) {
    output$test <- renderIhazr(
      ihazr(time=pbc5[,1], status=pbc5[,2], marker=pbc5[,3:7])
      )
  }
)

Example 4 -- mgus2
library(survival)
library(dplyr)
mgusN <- mgus2 \%>\%
  mutate(time = futime/12)
ihazr(time=mgusN$time, status=mgusN$death, marker=mgusN[,c(2:6,8)], buttons = F, binMax=8)
}
