% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihazr.R
\name{ihazr}
\alias{ihazr}
\title{ihazr}
\usage{
ihazr(time, status, marker, width = NULL, height = NULL,
  buttons = TRUE)
}
\arguments{
\item{time}{A numerical vector of time data.}

\item{status}{A binary vector representing event or no event.}

\item{marker}{A collection of numerical variables of interest for survival
analysis.}

\item{width}{The width of the html widget. The default is NULL, which results
in intelligent automatic sizing based on the widget's container.}

\item{height}{The height of the html widget. The default is NULL, which
results in intelligent automatic sizing based on the widget's container.}
}
\value{
ihazr of data values \code{time}, \code{status}, and \code{marker}
}
\description{
\code{ihazr} interactively updates a hazard function for survival anaylisis.
This function creates an HTML widget for survival data. The widget allows the
user to choose subsections of their data and plot the nonparametric kernel
estimated hazard function and cumulative hazard function.
}
\examples{
Example 1 - simulated data
time_t <- runif(50, 0, 10)
status_t <- rbinom(50, 1, .7)
age_t <- runif(50, 20, 80)
ihazr(time_t, status_t, age_t)

Example 2 - with survival data
library(survival)
library(dplyr)
pbc5 <- pbc \%>\%
  slice(1:312) \%>\%
  select(time, status, age, edema, bili, albumin, protime) \%>\%
  mutate(time = time/365, status = (status==2)*1, bili = log(bili),
    protime = log(protime))
ihazr(time=pbc5[,1], status=pbc5[,2], marker=pbc5[,3:7])

Example 3 - in a Shiny app
library(shiny)
library(survival)
library(dplyr)
pbc5 <- pbc \%>\% slice(1:312) \%>\%
      select(time, status, age, edema, bili, albumin, protime) \%>\%
      mutate(time = time/365, status = (status==2)*1, bili = log(bili),
        protime = log(protime))
shinyApp(
  ui = ihazrOutput("test"),
  server = function(input, output) {
    output$test <- renderIhazr(
      ihazr(time=pbc5[,1], status=pbc5[,2], marker=pbc5[,3:7])
      )
  }
)
}
